Adding CellulatBonding. Modified Board to get actionsOnMinions to select one Modified Incorp and MakeContact to be "copyable" and to call the super class Modified the super class to not actually push the action onto the minion if the action != this aka it is the hidden back class for cellular bonding.
The way that it works is that cellular bonding copies a class that is on a minion. Copying a class means creating a new one. It then sets that classes currentOwner to the owner of the cellular bonding. All of the calls (so far) are then forwarded to their hidden class if Cell Bonding receives it.

The hidden class doesn't actually get added to the minion because action != this in the super class call it makes play(minion,action).
The only reference to the hidden class is in the Cellular Bonding and it is destroyed when cell bonding is.

NOTE: I will have to add discard eventually....