Basic Additions to make the game playable
There are a couple of things done here. I added the ability to select a minionStruct option if there are multiple. I'm not 100% sure if I don't want to templitize the select "blah" methods that are now building up. There reason against it is to make attaching to the UI easier.

I'm also added removeCard from hand becase after we were playing cards it wasn't being properly removed.