Added BactaTheFuture and Finished the ShapeShifter Deck.
Bacta the future has an fPlay because it requires a minion to be destory. It can try to destroy minions that can't be destoryed so the only requirement is there is a minion on the board.

The way it works is you call destory, then the owner of the card you called destory on will get an extra minion.

If you destroy your own minion you don't need to do anything else special as the extra minion will be picked up in the playLoop.

If it isn't your turn you need to call playMinionOutOfTurn() which allows you to get at your playLoop again without doing all of the other actions associated with taking a turn.

The playLoop is the main logic that allows players to keep going while they have minions/actions to play if they want to (aka they don't select NULL) This was factored out of takeTurn() to allow playMinionsOutOfTurn to have access to it.