Changing move in card to not need reference to oldBase (Also, bug fix)
The thought process behind this change is that any card that can move (aka is played on a base) has a reference to the base it is currently played on so we don't need to be passing around a reference to that same base... just a way of introducing bugs.

The BactaTheFuture change is because I forgot to return true (or false) if there are minions in play to destroy. Whether or not it is true or false depends on the super class (where logic around actions remaining etc. etc. live)