Added fMovable also Mods to FirstMate
Need to check if something is moveable before actually moving it in things like PirateKing and FirstMate because of plant cards, like entangle.

Buccaneer doesn't need to be modified because he isn't destroyed even if he can't move.

FirstMate was modified to be discarded if the owner choose to not select a base.