Add SeaDogsAction
"Name a faction. Move all other player's minions of that faction from one base to another"

Added a way to get the faction from the Card
Added a selectFaction to the player (these selects are getting a little crazy lol)
Added a "minionsOfFaction" to base.... don't know if there are other cards that need it but this is where it should live

Last Check-in added the factions.