Added General Ivan
"Ongoing: Your minions cannot be destroyed"

I added a destroyableCounter for the player to store this information. There will likely need to be another one for bases.

I also added discard to the CopyCat so that it will play properly with minions that override the  General

In order to do that I changed discard to be able to take a MinionCard.

Also added discardFromHand because discard is really only once it is in play but some minions (aka Spectre still need to do things)