Added PolarCommando + smaller fixes
"Ongoing: If this is your only minion on this base, it has + 2 power and cannot be destroyed"

Added fDestroyable to CopyCat. Also, created fDestroyable(Minion) so that the CopyCat can send itself to other minions like it does for everything else :/ (still think it is an okay design)

Also added Bear Cavalry to DeckConstructor and Minions.

Also jokes (c++) lol

Fixed a spelling mistake with Destroy != destory