General Update and Fixes (I'll try to break it up and explain everything)
=======================
Fit and Finish to MinionCard::move
====================
The Minion Card tells the old base to remove it and the new base to "move" it (which is different than "playing" it. The new base will "set itself" on the card that it is adding.

removeMinion isn't really destroying it (aka when it moves). Therefore, it should have all its actions removed. Let the card remove its own actions when it knows its going to the discard.

In order to know if it is movable, a card will need to know where it is moving to and its current base. Thing could possibly prevent a card from moving to it?

===================
Polish on Base::destroyMinion
===================
The purpose of this method is so other bases can override it if they do special things (like giving people a vp). Really all this method is doing is telling the minion to go to the discard (which will then call removeMinion (possibly See Jumper)

=======================
Implement MinionCard::discard
=======================
Note: The MinionCard super class does all the heavy lifting. Subclasses will do special actions then either call the heavy lift or not (Jumpers for example may not)

To "discard" a card, first tell the base to remove it.  Then remove the reference to the base. Then remove all actions on the minion. Then add it to its owners discard pile (which was added)