Add Tenacious Z and Modified Spectre to play from discards
For both Tenacious Z and Spectre we needed a method to see if they are in the discard pile. Since the only way they can be played is if they are in the hand or the discards. It is easier/faster to say, if they aren't in the hand then they are in the discards.

This was added to fPlay of the Spectre.
Also added global (boo) bool playedTenaciousZ which marks whether or not the TenaciousZ has been played that turn. only zombies decks will ever modify that variable outside of begin and end turn.