Added Grave Robbing
"Place a card from your discard pile into your hand"

Also added discards method to player to get this discard pile.

Modified removeFromDiscard to return a reference to what was removed and fixed Across the divide to use that reference. Otherwise you would keep adding the same "card" back to the hand (and not remove it from the discards) which is bad ... just bad