Fixes
Changing Across the divide to use multiSelection. This required adding the method to player to get all cards of name x from discards. Changed from optional to not because you have to pick a card name. And then can select none of them.

DeadRiseAction now uses multiselect as well to selection cards from the current owners hand to discard, then using that size to determine the minion from the discard pile. Needed to add hand() method to Player to get the hand vector.

Switched the model from update to pull. When bases/minions want there power they pull from there action cards that return their value.

Fixed up some TODOs around base selection and moving in ActionCard.cpp

Changed shuffleMinionInDeck to not actually remove the minion from the base. It is the minions responsibility to do that. Also added it to shuffle the minion into the deck lol.