Added Aliens (as a deck) in addition to some Alien Minion Cards
Added Aliens as a faction to the global and to the deckConstructor

Also added:
Invader
Collector

Also refactored logic for parsing minions via power to the MinionCard object. (Began process of refactoring all the code to hit that)

Invader is simple and adds no new functions.
Collector adds returnToOwner() to the MinionCard (to be decided if this will be a "Card" method (most likely yes))